Import script no longer overwrites everything.

+++ scripts/nbl-to-ref.js
@@ -1,23 +1,25 @@
 // Convert NodeBox Live's /data/core/g/project.json to a list of reference documentation.
 // Execute:
 //    node scripts/nbl-to-ref.js <path to nodebox live>
+//
+// Use --overwrite to overwrite existing reference files. Otherwise they will be left alone.
 
 'use strict';
 
 var fs = require('fs');
 
-var nblPath = process.argv[2];
-console.assert(nblPath);
-var projectFile = nblPath + '/data/core/g/project.json';
+var projectFile = process.argv[2];
+console.assert(projectFile, 'Usage: node scripts/nbl-to-ref.js <path-to-project.json>');
+console.assert(fs.existsSync(projectFile), 'File project.json does not exist.');
+
+var overwrite = process.length == 4 && process.argv[3] === '--overwrite';
 
 console.assert(fs.existsSync('ref'), 'Execute this script in the g.js project directory.');
 
 var project = JSON.parse(fs.readFileSync(projectFile));
-console.log(project.id);
 
 for (var i = 0; i < project.functions.length; i++) {
     var fn = project.functions[i];
-    console.log(fn.name);
     var ref = fn.ref;
     ref = ref.replace(/\(guide:vector-graphics\)/, '(/guide/vector.html)');
     ref = ref.replace(/\(guide:(.*)\)/g, '(/guide/$1.html)');
@@ -31,6 +33,9 @@ for (var i = 0; i < project.functions.length; i++) {
     md += '---\n'
     md += ref;
     var functionFile = 'ref/' + fn.name + '.md';
-    fs.writeFileSync(functionFile, md);
+    if ((!fs.existsSync(functionFile)) || overwrite) {
+        console.log(fn.name);
+        fs.writeFileSync(functionFile, md);
+    }
 }
